requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic<2.0.0
requests==2.31.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langgraph==0.0.20
nltk==3.8.1
rapidfuzz==3.5.2
python-Levenshtein==0.23.0

